% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.sum.ess.R
\name{cor2cov}
\alias{cor2cov}
\title{convert correlation to covariance matrix}
\usage{
cor2cov(r2.mat, v.vec)
}
\arguments{
\item{r2.mat}{correlation matrix}

\item{v.vec}{variance vector}
}
\value{
covariance matrix
}
\description{
convert correlation to covariance matrix
}
