% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.sum.ess.R
\name{fast.lasso.sum.ess}
\alias{fast.lasso.sum.ess}
\title{Lasso calculation for summary statistics}
\usage{
fast.lasso.sum.ess(b.vec, s.vec, r2.mat, n, group, lambda, alpha,
  init.vec = NULL, max.iter = 100)
}
\arguments{
\item{b.vec}{genetic effect estimates}

\item{s.vec}{standard error of genetic effect estimates;}

\item{r2.mat}{residual errors;}

\item{n}{sample size;}

\item{group}{1-based group indicator}

\item{lambda}{l1 penalty parameter}

\item{alpha}{l2 penalty parameter}

\item{init.vec}{inital guess}

\item{max.iter}{maximum iteration}
}
\value{
a list: `beta` is the fitted regression coefficient vector, and `iteration` is the actual iteration.
}
\description{
Lasso calculation for summary statistics
}
