% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_lasso_sum_ess}
\alias{fast_lasso_sum_ess}
\title{Rewrite lasso.sum.ess}
\usage{
fast_lasso_sum_ess(bVec, sVec, r2Mat, n, group, lambda, alpha)
}
\arguments{
\item{bVec}{genetic effect estimates}

\item{sVec}{standard error of genetic effect estimates;}

\item{r2Mat}{residual errors;}

\item{n}{sample size;}

\item{group}{1-based group indicator}

\item{lambda}{l1 penalty parameter}

\item{alpha}{l2 penalty parameter}
}
\value{
fitted regression coefficient vector
}
\description{
Rewrite lasso.sum.ess
}
