% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.sum.ess.R
\name{lasso.sum.ess}
\alias{lasso.sum.ess}
\title{Lasso calculation for summary statistics}
\usage{
lasso.sum.ess(b.vec, s.vec, r2.mat, n, group, lambda, alpha)
}
\arguments{
\item{b.vec}{genetic effect estimates}

\item{s.vec}{standard error of genetic effect estimates;}

\item{r2.mat}{residual errors;}

\item{n}{sample size;}

\item{group}{1-based group indicator}

\item{lambda}{l1 penalty parameter}

\item{alpha}{l2 penalty parameter}
}
\value{
fitted regression coefficient vector
}
\description{
Lasso calculation for summary statistics
}
